Homebrew targets: use numeric types for controls values

+++ src/editor/editor.cpp
@@ -2063,7 +2063,7 @@ void GetEditorControls()
     {
 #ifdef __3DS__
         int win_x, win_y;
-        XRender::mapFromScreen(SharedCursor.X, SharedCursor.Y, &win_x, &win_y);
+        XRender::mapFromScreen((int)SharedCursor.X, (int)SharedCursor.Y, &win_x, &win_y);
 #endif
 
         editorScreen.active = !editorScreen.active;
